+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 * Optimize `image.allbut` in case where the new image attribute is
   thrown away. The image can be removed alltogether.
 
+* Optimize `transform.allbut`. The allbut can always be pushed down
+  the tree.
+
 # 0.18.0 - 2021/03/12
 
 * Default Relation#type is provided, that returns Bmg::Type::ANY